split http stuff into its own
